requirements.txt
django>=5.2
djangorestframework>=3.15
django-filter
djangorestframework-simplejwt
coverage
drf_yasg
drf-spectacular 
drf-spectacular-sidecar